test-engineering/contract-tests/partner/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.4.1
    # via uvicorn
click==8.0.1
    # via uvicorn
fastapi==0.68.2
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
httptools==0.4.0
    # via uvicorn
pydantic==1.8.2
    # via fastapi
python-dotenv==0.17.1
    # via uvicorn
pyyaml==6.0
    # via
    #   -r requirements.in
    #   uvicorn
starlette==0.14.2
    # via fastapi
typing-extensions==3.10.0.0
    # via pydantic
uvicorn[standard]==0.17.6
    # via -r requirements.in
uvloop==0.15.2
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==10.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
